Координатор ИИ. Было сделано следующее: - Доабвлена возможность смотреть историю чата - Добавлена кнопка "Отмена" при обработке запроса - Логирование истории чата идет теперь параллельно в БД SQL - Добавлена обратная связь по ответу ИИ, которая записывается в БД - Теперь можно прикреплять текстовые файлы (в том числе word) и картинке (в том числе через буфер обмена) - Исправлена ошибка с отменой запроса - Доабвлена пагинация как внутри команд так и в конечном ответе - Доабвлена возможность смотреть историю чата

+++ KPLN_CoordiantorAI/ExternalAIModel/ChatLogger.cs
@@ -228,10 +228,10 @@ namespace KPLN_CoordiantorAI.ExternalModel
         public void LogWithTokens(
             string question,
             string answer,
-            int cacheHitTokens,
-            int cacheMissTokens,
-            int completionTokens,
-            int totalTokens,
+            long cacheHitTokens,
+            long cacheMissTokens,
+            long completionTokens,
+            long totalTokens,
             double usdToRubRate = USD_TO_RUB_RATE)
         {
             DateTime timestamp = DateTime.Now;
@@ -261,10 +261,10 @@ namespace KPLN_CoordiantorAI.ExternalModel
         public void LogWithTokens(
             string question,
             string answer,
-            int cacheHitTokens,
-            int cacheMissTokens,
-            int completionTokens,
-            int totalTokens,
+            long cacheHitTokens,
+            long cacheMissTokens,
+            long completionTokens,
+            long totalTokens,
             DateTime requestTime,
             DateTime responseTime,
             string revitModelName,
